add unit-test and fix autorization and route

+++ app/Http/Controllers/API/AttachmentController.php
@@ -11,6 +11,7 @@ class AttachmentController extends BaseController
 {
     protected $attachmentService;
     public function __construct(AttachmentService $attachmentService) {
+        $this->middleware('auth:api')->except(['index', 'show', 'getByPostId']);
         $this->attachmentService = $attachmentService;
     }
 
@@ -21,6 +22,10 @@ class AttachmentController extends BaseController
 
     public function store(Request $request)
     {
+        if (!$this->middleware('auth:api')->passes()) {
+            return $this->sendError('Unauthorized', [], 401);
+        }
+
         $input = $request->all();
         $file = $request->file('attachmentfile');
         $result = $this->attachmentService->store($input, $file);
@@ -41,6 +46,10 @@ class AttachmentController extends BaseController
 
     public function updateById(Request $request, $id)
     {
+        if (!$this->middleware('auth:api')->passes()) {
+            return $this->sendError('Unauthorized', [], 401);
+        }
+
         $file = $request->file('attachmentfile');
         $result = $this->attachmentService->updateById($file, $id);
         if (is_string($result)) {
@@ -51,6 +60,10 @@ class AttachmentController extends BaseController
 
     public function destroy(Attachment $attachment)
     {
+        if (!$this->middleware('auth:api')->passes()) {
+            return $this->sendError('Unauthorized', [], 401);
+        }
+        
         $result = $this->attachmentService->destroy($attachment);
         return $this->sendResponse($result->toArray(), 'Attachment deleted successfully.');
     }
